It compiles!
