+++ meson.build
@@ -45,6 +45,7 @@ sources = files([
     'src/Configuration.cpp',
     'src/FtlClient.cpp',
     'src/FtlControlConnection.cpp',
+    'src/FtlMediaConnection.cpp',
     'src/FtlServer.cpp',
     'src/FtlStream.cpp',
     'src/JanusFtl.cpp',
