+++ src/ConnectionTransports/ConnectionTransport.h
@@ -37,32 +37,20 @@ public:
 
     /**
      * @brief
-     *  Starts the connection transport in a new thread.
-     *  This function should block until the transport is ready to receive bytes.
+     *  Shuts down the connection.
+     *  This function should block until the underlying transport/socket has been closed.
      */
-    virtual Result<void> StartAsync() = 0;
+    virtual void Stop() = 0;
 
     /**
      * @brief
-     *  Shuts down the connection.
-     *  This function should block until the underlying transport/socket has been closed, unless
-     *  noBlock has been set.
+        Read a set of bytes from the transport into the given buffer.
+        Will timeout if there is nothing to read and return zero bytes.
      */
-    virtual void Stop(bool noBlock = false) = 0;
+    virtual Result<ssize_t> Read(std::vector<std::byte>& bytes) = 0;
 
     /**
      * @brief Write a set of bytes to the transport
      */
-    virtual void Write(const std::vector<std::byte>& bytes) = 0;
-
-    /**
-     * @brief Sets the callback that will fire when this connection has been closed.
-     */
-    virtual void SetOnConnectionClosed(std::function<void(void)> onConnectionClosed) = 0;
-
-    /**
-     * @brief Sets the callback that will fire when this connection has received incoming data.
-     */
-    virtual void SetOnBytesReceived(
-        std::function<void(const std::vector<std::byte>&)> onBytesReceived) = 0;
+    virtual Result<void> Write(const std::vector<std::byte>& bytes) = 0;
 };
