+++ src/FtlStream.h
@@ -9,6 +9,7 @@
  */
 #pragma once
 
+#include "FtlMediaConnection.h"
 #include "Utilities/FtlTypes.h"
 #include "Utilities/Result.h"
 
@@ -33,34 +34,19 @@ class FtlStream
 public:
     /* Public types */
     using ClosedCallback = std::function<void(FtlStream*)>;
-    using RtpPacketCallback = std::function<void(const std::vector<std::byte>&)>;
-    struct FtlStreamStats
-    {
-        time_t StartTime;
-        uint32_t DurationSeconds;
-        uint32_t RollingAverageBitrateBps;
-        uint32_t PacketsReceived;
-        uint32_t PacketsNacked;
-        uint32_t PacketsLost;
-    };
-    struct FtlKeyframe
-    {
-        VideoCodecKind Codec;
-        std::list<std::vector<std::byte>> Packets;
-    };
 
     /* Constructor/Destructor */
     FtlStream(
         std::shared_ptr<FtlControlConnection> controlConnection,
-        std::unique_ptr<ConnectionTransport> mediaTransport,
-        const MediaMetadata mediaMetadata,
-        const ftl_stream_id_t streamId,
-        const ClosedCallback onClosed,
-        const RtpPacketCallback onRtpPacket,
-        const bool nackLostPackets = true);
+        ftl_stream_id_t streamId,
+        const ClosedCallback onClosed);
 
     /* Public methods */
-    Result<void> StartAsync(uint16_t mediaPort);
+    Result<void> StartMediaConnection(
+        std::unique_ptr<ConnectionTransport> mediaTransport,
+        const MediaMetadata mediaMetadata,
+        const FtlMediaConnection::RtpPacketCallback onRtpPacket
+    );
     void Stop();
     void ControlConnectionStopped(FtlControlConnection* controlConnection);
 
@@ -71,63 +57,14 @@ public:
     FtlKeyframe GetKeyframe();
 
 private:
-    /* Private types */
-    struct SsrcData
-    {
-        uint32_t PacketsReceived = 0;
-        uint32_t PacketsNacked = 0;
-        uint32_t PacketsLost = 0;
-        size_t PacketsSinceLastMissedSequence = 0;
-        std::list<std::vector<std::byte>> CircularPacketBuffer;
-        std::map<std::chrono::time_point<std::chrono::steady_clock>, uint16_t> 
-            RollingBytesReceivedByTime;
-        std::set<rtp_sequence_num_t> NackQueue;
-        std::set<rtp_sequence_num_t> NackedSequences;
-        std::list<std::vector<std::byte>> CurrentKeyframePackets;
-        std::list<std::vector<std::byte>> PendingKeyframePackets;
-    };
-
-    /* Constants */
-    static constexpr uint64_t            MICROSECONDS_PER_SECOND        = 1000000;
-    static constexpr float               MICROSECONDS_PER_MILLISECOND   = 1000.0f;
-    static constexpr rtp_payload_type_t  FTL_PAYLOAD_TYPE_SENDER_REPORT = 200;
-    static constexpr rtp_payload_type_t  FTL_PAYLOAD_TYPE_PING          = 250;
-    static constexpr size_t              PACKET_BUFFER_SIZE             = 128;
-    static constexpr size_t              MAX_PACKETS_BEFORE_NACK        = 16;
-    static constexpr size_t              NACK_TIMEOUT_SEQUENCE_DELTA    = 128;
-    static constexpr uint32_t            ROLLING_SIZE_AVERAGE_MS        = 2000;
-
     /* Private members */
     const std::shared_ptr<FtlControlConnection> controlConnection;
-    const std::unique_ptr<ConnectionTransport> mediaTransport;
-    const MediaMetadata mediaMetadata;
+    std::unique_ptr<FtlMediaConnection> mediaConnection;
     const ftl_stream_id_t streamId;
     const ClosedCallback onClosed;
-    const RtpPacketCallback onRtpPacket;
-    // Stream data
-    std::shared_mutex dataMutex;
-    time_t startTime { 0 };
-    std::chrono::time_point<std::chrono::steady_clock> steadyStartTime;
-    std::unordered_map<rtp_ssrc_t, SsrcData> ssrcData;
+    std::mutex mutex;
 
     /* Private methods */
-    void mediaBytesReceived(const std::vector<std::byte>& bytes);
-    void mediaConnectionClosed();
-    // Packet processing
-    std::set<rtp_sequence_num_t> insertPacketInSequenceOrder(
-        std::list<std::vector<std::byte>>& packetList, const std::vector<std::byte>& packet);
-    void processRtpPacket(const std::vector<std::byte>& rtpPacket);
-    void processRtpPacketSequencing(const std::vector<std::byte>& rtpPacket,
-        const std::unique_lock<std::shared_mutex>& dataLock);
-    void processRtpPacketKeyframe(const std::vector<std::byte>& rtpPacket,
-        const std::unique_lock<std::shared_mutex>& dataLock);
-    void processRtpH264PacketKeyframe(const std::vector<std::byte>& rtpPacket,
-        const std::unique_lock<std::shared_mutex>& dataLock);
-    bool isSequenceNewer(rtp_sequence_num_t newSeq, rtp_sequence_num_t oldSeq,
-        size_t margin = PACKET_BUFFER_SIZE);
-    void processNacks(const rtp_ssrc_t ssrc, const std::unique_lock<std::shared_mutex>& dataLock);
-    void processAudioVideoRtpPacket(const std::vector<std::byte>& rtpPacket,
-        std::unique_lock<std::shared_mutex>& dataLock);
-    void handlePing(const std::vector<std::byte>& rtpPacket);
-    void handleSenderReport(const std::vector<std::byte>& rtpPacket);
+    void onControlConnectionClosed();
+    void onMediaConnectionClosed();
 };
